pyproject.toml
[tool.poetry]
name = "notification_py"
version = "1.0.4"
description = "A simple package to send notifications on Discord, Slack and Email."
authors = ["Pranav Soni <pranav.bhawan@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
repository = "https://github.com/ps428/notification-py/"
packages = [{include = "notification_py"}]


[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.6.4"
aiohttp = "^3.9.3"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
aiosmtplib = "^3.0.1"
print-position = "^2.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^6.1.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



